% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuminatoR.r
\name{compute.statistics}
\alias{compute.statistics}
\title{Compute the statistics for classification}
\usage{
compute.statistics(data, prediction = FALSE, min.amplitude = 30, min.dt = 6)
}
\arguments{
\item{data}{data.frame wite a columns pressure}

\item{prediction}{if FALSE, the obeserved activities are added to the output}

\item{min.dt}{minimal time difference between peaks}

\item{minimal.amplitude}{minimal amplitude to count as peak}
}
\value{
a list with the statistics of the peaks, and the time points
}
\description{
Compute the statistics for classification
}
