% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuminatoR.r
\name{identify.peaks}
\alias{identify.peaks}
\title{Identify peaks in time series}
\usage{
\method{identify}{peaks}(pressure, min.amplitude = 30, min.dt = 6)
}
\arguments{
\item{pressure}{time series of pressure data}

\item{min.amplitude}{minimal amplitude to count as peak}

\item{min.dt}{minimal time difference between peaks}
}
\value{
a vector of the indicies of the peaks
}
\description{
Identify peaks in time series
}
