% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuminatoR.r
\name{train.RF}
\alias{train.RF}
\title{Train Random Forest}
\usage{
train.RF(data, min.amplitude = 30, min.dt = 6, ...)
}
\arguments{
\item{data}{A \code{data.frame} wite a columns \code{time}, \code{pressure}, and \code{activity}}

\item{min.dt}{minimal time difference between peaks}

\item{...}{arguments passed to \code{randomForest}}

\item{minimal.amplitude}{minimal amplitude to count as peak}
}
\value{
list of with trained randomForest and the parameters used.
}
\description{
Train Random Forest
}
\author{
Andreas Scheidegger
}
