% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuminatoR.r
\name{identify.peaks}
\alias{identify.peaks}
\title{Identify indices of peaks}
\usage{
\method{identify}{peaks}(pressure, min.amplitude = 30, min.dt = 6)
}
\arguments{
\item{min.dt}{minimal time difference between peaks}

\item{minimal.amplitude}{minimal amplitude to count as peak}
}
\value{
a vector of the indicies of the peaks
}
\description{
Identify indices of peaks
}
